Matching semaphores to new structure
Finally getting the semaphores implemented with the correct functions now that consolidation is finished. Should take remaining images into account so that we only need to call the for loop, and not have a remaining images loop. Neither half of the producer/consumer loop can proceed without the previous section finishing